Deconstruction of d3drender for textures
