+++ clientd3d/makefile
@@ -127,6 +127,7 @@ OBJS = \
         $(OUTDIR)\d3dparticle.obj \
         $(OUTDIR)\d3drender.obj \
         $(OUTDIR)\d3drender_objects.obj \
+        $(OUTDIR)\d3drender_textures.obj \
         $(OUTDIR)\d3drender_world.obj \
         $(OUTDIR)\matrix.obj \
         $(OUTDIR)\xform.obj \
